+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Changed
+- `ProductConfigManager::from_yaml_file` now takes a `&Path` as well as `&str` ([#43])
+
+[#43]: https://github.com/stackabletech/product-config/pull/43
 
 ## [0.2.0] - 2021-11-05